requirements.txt
numpy>=1.26
pandas>=2.2
pyarrow>=17.0
fastf1>=3.4
requests>=2.32
scikit-learn>=1.5
scipy>=1.13
cvxpy>=1.5
shapely>=2.0
plotly>=5.24
streamlit>=1.39
pyyaml>=6.0
tqdm>=4.66
